+++ doc/changelog.rst
@@ -43,6 +43,7 @@ Core
 Admin
 ~~~~~
 
+- Add middleware to select and set the current shop in the request
 - Breaking change: add optional shop parameter in `get_model_url` method of admin module
 - Add middleware to select the active admin shop using session
 - Only categories, orders, products, contacts, campaigns and services for the active shop are displayed in the admin